+++ types/global.d.ts
@@ -16,3 +16,5 @@ declare const __IS_DYNAMIC_CUSTOM_BUNDLE__: boolean;
 declare const __IS_LEGACY_BUILD__: boolean;
 
 declare const __BUNDLED_PLUGINS_LIST__: string | undefined;
+
+declare const __LOCK_DEPS_VERSION__: boolean;