+++ packages/analytics-js-integrations/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.12.2](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-integrations@3.12.1...@rudderstack/analytics-js-integrations@3.12.2) (2025-01-24)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.15.0`
 ## [3.12.1](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-integrations@3.12.0...@rudderstack/analytics-js-integrations@3.12.1) (2025-01-22)